backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.14.0
astroid==3.3.5
attrs==25.4.0
auth0-python==4.7.2
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
cryptography==43.0.3
dill==0.4.0
ecdsa==0.19.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
idna==3.11
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.7.0
platformdirs==4.5.1
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.21.0
PyJWT==2.10.1
pylint==3.3.2
python-dotenv==1.2.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
tomlkit==0.13.3
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.3
yarl==1.22.0
pylint==3.3.2
isort==5.13.2
gunicorn==20.1.0
psycopg2-binary==2.9.9